+++ custom_components/unifi_unas_rest/const.py
@@ -13,6 +13,7 @@ PLATFORMS: Final = [
     Platform.BUTTON,
     Platform.SELECT,
     Platform.SENSOR,
+    Platform.SWITCH,
     Platform.UPDATE,
 ]